Favicons, nav in header
Favicons were added to the header of every generated page. They are
stored in `/icon`. Additionally, the navigation buttons were moved into
the header as well. A positive side effect of this is that the top
banner always says 'Pope High School Computing Club', with an actual
title for the content as a later <h1>. The <title> is still correct, as
it is taken from the generated HTML fragment, before the header is ever
prepended to it.